+++ fiddles/angular2-cli/fiddle-0000-Template/src/app/fiddle.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  moduleId: module.id,
+  selector: 'fiddle-app',
+  templateUrl: 'fiddle.component.html',
+  styleUrls: ['fiddle.component.css']
+})
+export class FiddleAppComponent {
+  title = 'fiddle works!';
+}