+++ fiddles/angular2-cli/fiddle-0000-Template/src/main.ts
@@ -1,9 +1,9 @@
 import { bootstrap } from '@angular/platform-browser-dynamic';
 import { enableProdMode } from '@angular/core';
-import { Fiddle0000TemplateAppComponent, environment } from './app/';
+import { FiddleAppComponent, environment } from './app/';
 
 if (environment.production) {
   enableProdMode();
 }
 
-bootstrap(Fiddle0000TemplateAppComponent);
+bootstrap(FiddleAppComponent);