+++ fiddles/angular2/fiddle-0037-ModalPopUp/app/app.component.ts
@@ -14,10 +14,4 @@ import {GlobalService} from './global/global.service';
 export class AppComponent {
 
 
-
-    constructor() {
-
-    }
-
-
 }