+++ fiddles/angular2/fiddle-0038-DragAndDrop/app/app.component.ts
@@ -2,6 +2,10 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'my-app',
-  template: '<h1>My First Angular 2 App</h1>'
+  template: '<h1>Under Construction ...</h1>'
 })
-export class AppComponent { }
+export class AppComponent {
+
+
+
+}