+++ fiddles/angular2/fiddle-0038-DragAndDrop/app/util/offset.ts
@@ -1,4 +1,7 @@
+import {Injectable} from '@angular/core';
+import Injectable = ng.Injectable;
 
+@Injectable()
 export interface OffsetInterface {
     x:number;
     y:number;