+++ fiddles/angular2/fiddle-0038-DragAndDrop/app/util/util.ts
@@ -1,11 +1,9 @@
 
 export class Util {
-
     static get radianToDegreeConstant():number {
         return 180 / Math.PI;
     }
-
-    static constrain(num:number, min:number, max:number):number {
+    static constrain(num:any, min:number, max:number):number {
         let x = parseFloat(num);
         if (min === null) {
             min = num;
@@ -15,5 +13,4 @@ export class Util {
         }
         return (x < min) ? min : ((x > max) ? max : x);
     }
-
 }