+++ fiddles/bash/fiddle-0001-FunctionBasics/script.sh
@@ -11,7 +11,20 @@ function valueFunction() {
     z=`expr $1 + 3`;
     echo ${z};
 }
-controlFunction;
-controlFunction "Bob";
-rc=$(valueFunction 2;);
-echo "2 + 3 = ${rc}";
+
+function subDelimStr() {
+    _str=$1;
+    _delm=$2;
+    _pos=$3;
+    IFS=${_delm} read -r -a array <<< "${_str}";
+    _arrayLength=$(echo "${#array[@]}";);
+    if [[ ${_pos} -lt ${_arrayLength} ]]
+    then
+        echo ${array[${_pos}]};
+    else
+        echo "";
+    fi
+}
+
+
+subDelimStr "fiddle-0000-Template" "-" 2;