+++ fiddles/bash/fiddle-0090-Zmq/bin/_utils.sh
@@ -183,18 +183,4 @@ function stripLine() {
     fi
 }
 
-function subDelimStr() {
-    _str=$1;
-    _delm=$2;
-    _pos=$3;
-    IFS=${_delm} read -r -a array <<< "${_str}";
-    _arrayLength=$(echo "${#array[@]}";);
-    if [[ ${_pos} -lt ${_arrayLength} ]]
-    then
-        echo ${array[${_pos}]};
-    else
-        echo "";
-    fi
-}
-