+++ fiddles/bash/fiddle-0090-Zmq/bin/brew/_update.sh
@@ -1,0 +1,19 @@
+#!/usr/bin/env bash
+
+this=$(pwd;);
+
+function brewUpdate() {
+    groupLog "brewUpdate";
+    installed=$(isInstalled "brew";);
+    if [[ "${installed}" == "false" ]]
+    then
+        groupLog "${this}: home brew is not installed";
+        exit -1;
+    else
+        brew update;
+    fi
+}
+
+
+
+