+++ fiddles/bash/fiddle-0090-Zmq/bin/zmq/_uninstall.sh
@@ -1,0 +1,15 @@
+#!/usr/bin/env bash
+
+this=$0;
+
+function zmqUninstall() {
+    groupLog "brewUpdate";
+    installed=$(isInstalled "brew";);
+    if [[ "${installed}" == "false" ]]
+    then
+        groupLog "${this}: home brew is not installed";
+        exit -1;
+    else
+        brew uninstall --force zeromq || exit $?;
+    fi
+}